removed magic stuff
