+++ src/main.rs
@@ -7,7 +7,6 @@
 mod board_representation;
 mod chess_move;
 mod lookup;
-mod magic;
 mod util_macros;
 
 fn main() {
